feat/fix: implement intelligent fallback pricing based on product category and condition; fix createdAt serialization for frontend integration

+++ src/main/java/com/yann/smart_valuator_api/DTO/AiEstimationResult.java
@@ -3,6 +3,8 @@ package com.yann.smart_valuator_api.DTO;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class AiEstimationResult {
 
@@ -10,7 +12,7 @@ public class AiEstimationResult {
     private String description;
 
     @JsonProperty("estimatedPrice")
-    private Double estimatedPrice;
+    private BigDecimal estimatedPrice;
 
     @JsonProperty("verdict")
     private String verdict;
